#48: Move some more files
